requirements/dev.txt
-r base.txt

flake8==4.0.1
black==22.3.0
autoflake~=1.4
isort~=5.10.1
pylint~=2.12.2
pylint-django==2.5.2